Fix svg element creation in Html5
Use [Svg.svg] to create an svg embedded in a [Html5.elt]. This is
needed because in Javascript SVG nodes should be created using the
[createElementNS] method, while all other DOM elements are typically
created using the [createElement] method. This change is useful to fix
a bug in js_of_ocaml's [Tyxml_js] module (see
ocsigen/js_of_ocaml#276).