Adapt the PPX to upcoming Markup.ml 0.9.0
Markup.ml 0.9.0 adds a new signal, `Raw of string. When the HTML and
XML writers encounter it, they output the string payload verbatim. The
parsers never generate this signal.

Since the PPX only uses the HTML parser, it is not affected, except that
we have to add a case to ignore this signal (which the PPX will never
get).