doc:ci: fix symlink generation
The symbolic links were not deployed on 'gh-pages' because of a Travis
bug (see https://github.com/travis-ci/dpl/issues/912).

Manually adding the symlinks doesn't work either because html_of_wiki
generates *relative links* based on the position of the file during
the compilation. Hence these links, from a page with a different URL,
are dead.
For example, having:
1. tyxml/index.html linking to tyxml/4.3.0/manual/intro.html,
2. inside tyxml/4.3.0/manual/intro.html:
        <a href="functors.html">...</a>,
3. then, from the URL https://ocsigen.org/tyxml, the link above
   is dead, hence the real http redirection.