Port the 3 ppx's to `ppxlib`
Before, the 2 ppx rewriters `tyxml-ppx` and `tyxml-jsx` as well as the
internal `.ml`-file generator in /syntax/ were based on `omp`: the former
two on the `omp` driver and all of them on `Ast_408`. None of those will
exist in `omp 2.0.0` anymore.

This commit ports those 3 ppx's from the direct use of `omp` and tools
like `ppx_tools_versioned` to `ppxlib`.